add saturation to mixer

+++ include/mixer.h
@@ -26,18 +26,18 @@ typedef struct {
 
 typedef struct {
   output_type_t output_type[8];
-  int16_t F[8];
-  int16_t x[8];
-  int16_t y[8];
-  int16_t z[8];
+  int32_t F[8];
+  int32_t x[8];
+  int32_t y[8];
+  int32_t z[8];
 } mixer_t;
 
 extern command_t _command;
 
-extern int16_t _GPIO_outputs[8];
+extern int32_t _GPIO_outputs[8];
 extern output_type_t _GPIO_output_type[8];
 
-extern int16_t _outputs[8];
+extern int32_t _outputs[8];
 
 void init_mixing();
 void mix_output();
